refactor(terminal): delete the superseded CLI-side agent hook client
The console CLI dispatches every 'hook <name>' subcommand through
console-lifecycle.ts, which posts to the lock endpoint itself. The agent-api
copies — postConsoleAgentHook and the attention, auto-name and turn-state
runners — reached no production caller; their only importers were two tests,
and their payload shape ({reason}, {prompt}) is the older protocol the server
still tolerates alongside the {input} one the live path sends.

normalizeAttentionReason survives because agent.ts normalizes the inbound
reason with it. It moves next to AgentAttentionReason, and the reason list
becomes the single source the union derives from — the list and the union were
written out twice with nothing keeping them in sync. A focused test replaces
the two deleted suites and guards the invariant they covered: a
notification_type outside the list is dropped rather than forwarded.
